+++ lib/new_relic/telemetry_sdk/harvester.rb
@@ -1,0 +1,59 @@
+# encoding: utf-8
+# This file is distributed under New Relic's license terms.
+# See https://github.com/newrelic/newrelic-telemetry-sdk-ruby/blob/main/LICENSE for complete details.
+
+module NewRelic
+  module TelemetrySdk
+    class Harvester
+
+      attr_reader :interval
+
+      def initialize interval = 5
+        @interval = interval
+        @harvestables = {}
+        @shutdown = false
+        @lock = Mutex.new
+      end
+
+      def register name, buffer, client
+        @lock.synchronize do 
+          @harvestables[name] = {
+            buffer: buffer,
+            client: client
+          }
+        end
+      end
+
+      def lookup name 
+        @harvestables[name]
+      end
+
+      def start
+        @harvest_thread = Thread.new do
+          while !@shutdown do
+            sleep @interval
+            harvest
+          end
+        end
+      end
+
+      def stop
+        @shutdown = true
+        @harvest_thread.join
+      end
+
+      def harvest
+        @lock.synchronize do
+          @harvestables.each do |harvestable|
+            process_harvestable harvestable
+          end
+        end
+      end
+
+      def process_harvestable harvestable
+        harvestable[:client].report_batch harvestable[:buffer].flush
+      end
+
+    end
+  end
+end
